updated controllers22

+++ src/main/java/com/example/ourstoryapp/web/PictureController.java
@@ -4,6 +4,8 @@ import java.net.URI;
 
 import javax.validation.Valid;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -23,29 +25,46 @@ public class PictureController {
 
 	@Autowired
 	PictureRepository repository;
+    Logger logger = LogManager.getLogger(PictureController.class);
 
 	@GetMapping("/findAll")
 	public Iterable<Picture> findAll() {
+		logger.info("Find All Pictures");
 		return repository.findAll();
 	}
 
 	@PostMapping("/create")
 	public Picture create(@Valid @RequestBody Picture picture) {
+		logger.info("Cretae Picture");
 		return repository.save(picture);
 	}
 
 	@GetMapping("/findById/{id}")
 	public ResponseEntity<Picture> findById(@PathVariable(value = "id") URI pictureId) {
-		return repository.findById(pictureId).map(record -> ResponseEntity.ok().body(record))
-				.orElse(ResponseEntity.notFound().build());
+		if(repository.findById(pictureId).map(record -> ResponseEntity.ok().body(record)).isPresent()) {
+			logger.info("Find Pictrure By ID");
+			return repository.findById(pictureId).map(record -> ResponseEntity.ok().body(record))
+					.orElse(ResponseEntity.notFound().build());
+		}
+		else {
+			logger.info("Picture By ID Is Not Found");
+			return ResponseEntity.notFound().build();
+		}
 	}
 
 	@DeleteMapping("/delete/{id}")
 	public ResponseEntity<?> delete(@PathVariable("id") URI pictureId) {
-		return repository.findById(pictureId).map(record -> {
-			repository.deleteById(pictureId);
-			return ResponseEntity.ok().build();
-		}).orElse(ResponseEntity.notFound().build());
+		if(repository.findById(pictureId).map(record -> ResponseEntity.ok().body(record)).isPresent()) {
+			logger.info("Successfully Deleted Picture");
+			return repository.findById(pictureId).map(record -> {
+				repository.deleteById(pictureId);
+				return ResponseEntity.ok().build();
+			}).orElse(ResponseEntity.notFound().build());
+		}
+		else {
+			logger.info("Picture Is Not Existing");
+			return ResponseEntity.notFound().build();
+		}
 	}
 
 }
